Improve performance for larger models by not projecting vertices redundantly for each face.
Improve performance by not using matrix multiplication for the perspective projection (test case of two models with 414 triangles in total: 20.2 ms vs. 24.4 ms for our pre-rasterisation).

Word-align structs (just to be safe; devkitarm does not guarantee word alignment, which was a big performance snag when RasterPoint still had an additional bool field. The non-word alignment degraded performance *very significantly* (from about 9.5 to 16.3 ms in the example I've tested).

Change the default cube's winding order to counter-clockwise (as a workaround to be compatible with blender's .obj winding order; I think we should be able to fix backface culling to work for any winding order, as we have correct face normals for our models anyway, which we could just rotate according to the instance's yaw/roll/pitch).